requirements.txt
flask
openai
sentence-transformers
torch
numpy
pandas
PyMuPDF
tqdm
python-dotenv
flask-cors